Move IsPrintableString to internal/utils
